chore(rust): restore apple clippy gate
Close P2-CLIPPY-CLEANUP-1 by making the Apple Silicon clippy gate executable with cargo clippy --target aarch64-apple-darwin -- -D warnings.

Fix low-risk core/type lints directly, codify the engine/bin historical lint baseline at crate boundaries, and keep the allowed classes explicit so new lint categories still fail. Preserve NaN fail-closed semantics in external paper-order checks.

Verification:
- cargo clippy --target aarch64-apple-darwin -- -D warnings
- cargo test -p openclaw_core --lib
- cargo test -p openclaw_engine --lib
- git diff --check
- git diff --cached --check

Boundary: no runtime/deploy/rebuild/restart, no DB/auth/risk/order/trading mutation, no real Bybit call.

+++ rust/openclaw_core/src/guardian.rs
@@ -210,7 +210,7 @@ impl Guardian {
             risk_score += SCORE_DRAWDOWN_BREACH;
         }
 
-        let risk_score = (risk_score as f64).min(RISK_SCORE_CAP);
+        let risk_score = risk_score.min(RISK_SCORE_CAP);
 
         // Verdict logic
         let verdict = if reasons.iter().any(|r| {
